requirements.txt
langchain==0.1.17
langchain-core==0.1.53
langchain-community==0.0.38
langchain-groq==0.1.3

faiss-cpu
pymupdf
streamlit
python-dotenv
tiktoken
reportLab
pandas
matplotlib
plotly
requests
